+++ src/main/java/com/example/jomajomadelivery/account/oauth2/dto/response/NaverUserInfoResponse.java
@@ -1,0 +1,36 @@
+package com.example.jomajomadelivery.account.auth.dto.response;
+
+import com.example.jomajomadelivery.account.auth.oauth2.SocialProvider;
+
+import java.util.Map;
+
+public class NaverUserInfoResponse implements OAuth2UserInfo {
+
+    private final Map<String, Object> attribute;
+
+    private final SocialProvider socialProvider = SocialProvider.NAVER;
+
+    public NaverUserInfoResponse(Map<String, Object> attributes) {
+        this.attribute = (Map<String, Object>) attributes.get("response");
+    }
+
+    @Override
+    public SocialProvider getProvider() {
+        return socialProvider;
+    }
+
+    @Override
+    public String getProviderId() {
+        return attribute.get("id").toString();
+    }
+
+    @Override
+    public String getName() {
+        return attribute.get("name").toString();
+    }
+
+    @Override
+    public String getEmail() {
+        return attribute.get("email").toString();
+    }
+}
