+++ src/main/java/com/example/jomajomadelivery/account/oauth2/dto/response/OAuth2UserInfo.java
@@ -1,0 +1,14 @@
+package com.example.jomajomadelivery.account.auth.dto.response;
+
+import com.example.jomajomadelivery.account.auth.oauth2.SocialProvider;
+
+public interface OAuth2UserInfo {
+
+    SocialProvider getProvider();
+
+    String getProviderId();
+
+    String getName();
+
+    String getEmail();
+}
