+++ src/main/java/com/example/jomajomadelivery/account/oauth2/service/CustomOAuth2User.java
@@ -1,0 +1,34 @@
+package com.example.jomajomadelivery.account.auth.oauth2;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.Collection;
+import java.util.Map;
+
+@Getter
+@RequiredArgsConstructor
+public class CustomOAuth2User implements OAuth2User {
+
+    private final SocialProvider provider;
+    private final String providerId;
+    private final String email;
+    private final String name;
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return Map.of();
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+}
