--- src/main/java/com/example/jomajomadelivery/auth/controller/SocialAuthController.java
@@ -1,9 +1,0 @@
-package com.example.jomajomadelivery.auth.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class SocialAuthController {
-
-    // Todo: 소셜 로그인/회원가입 담당
-}
