--- src/main/java/com/example/jomajomadelivery/auth/dto/response/OAuth2UserInfo.java
@@ -1,14 +1,0 @@
-package com.example.jomajomadelivery.auth.dto.response;
-
-import com.example.jomajomadelivery.auth.oauth2.SocialProvider;
-
-public interface OAuth2UserInfo {
-
-    SocialProvider getProvider();
-
-    String getProviderId();
-
-    String getName();
-
-    String getEmail();
-}
