--- src/main/java/com/example/jomajomadelivery/auth/jwt/TokenProvider.java
@@ -1,69 +1,0 @@
-package com.example.jomajomadelivery.auth.jwt;
-
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-
-@Slf4j
-@Component
-public class TokenProvider {
-
-    private SecretKey key;
-    private static final long TOKEN_VALID_TIME = 1000L * 60 * 60; // 토큰 유효 시간: 1시간
-
-    public TokenProvider(@Value("${spring.jwt.secret-key}") String key) {
-        this.key = Keys.hmacShaKeyFor(key.getBytes());
-    }
-
-    public String createToken(String email, List<String> roles) {
-        HashMap<String, Object> claims = new HashMap<>();
-        claims.put("email", email);
-        claims.put("roles", roles);
-
-        return Jwts.builder()
-                .claims(claims)
-                .issuedAt(new Date(System.currentTimeMillis()))
-                .expiration(new Date(System.currentTimeMillis() + TOKEN_VALID_TIME))
-                .signWith(key)
-                .compact();
-    }
-
-    public String getEmail(String token) {
-        return Jwts.parser()
-                .verifyWith(key)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .get("email", String.class);
-    }
-
-    public List<String> getRole(String token) {
-        return  (List<String>) Jwts.parser()
-                .verifyWith(key)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .get("roles");
-    }
-
-    public Boolean isExpired(String token) {
-        try {
-            Jwts.parser()
-                    .verifyWith(key)
-                    .build()
-                    .parseSignedClaims(token);
-            return false;
-        } catch (JwtException e) {
-            log.info("token 만료");
-            return true;
-        }
-    }
-}
