--- src/main/java/com/example/jomajomadelivery/auth/oauth2/CustomOAuth2User.java
@@ -1,34 +1,0 @@
-package com.example.jomajomadelivery.auth.oauth2;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-
-@RequiredArgsConstructor
-public class CustomOAuth2User implements OAuth2User {
-
-    private final String email;
-    private final List<String> roles;
-
-    @Override
-    public Map<String, Object> getAttributes() {
-        return Map.of();
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return roles.stream()
-                .map(SimpleGrantedAuthority::new)
-                .toList();
-    }
-
-    @Override
-    public String getName() {
-        return email;
-    }
-}
