--- src/main/java/com/example/jomajomadelivery/auth/oauth2/handler/CustomAuthenticationFailureHandler.java
@@ -1,28 +1,0 @@
-package com.example.jomajomadelivery.auth.oauth2.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Slf4j
-@Component
-public class CustomAuthenticationFailureHandler implements AuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException {
-        if (exception instanceof OAuth2AuthenticationException ex && ex.getError().getErrorCode().equals("redirect_signup")) {
-            log.info("Social Login Fail - Redirecting to /");
-
-            response.sendRedirect("/login");
-            return;
-        }
-
-        response.sendRedirect("/login?error=true");
-    }
-}
