+++ src/main/java/com/example/jomajomadelivery/user/util/BCryptUtil.java
@@ -1,0 +1,16 @@
+package com.example.jomajomadelivery.account.auth.util;
+
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+public class BCryptUtil {
+
+    private static final BCryptPasswordEncoder ENCODER = new BCryptPasswordEncoder();
+
+    public static String encrypt(String rawPassword) {
+        return ENCODER.encode(rawPassword);
+    }
+
+    public static boolean matches(String rawPassword, String encryptedPassword) {
+        return ENCODER.matches(rawPassword, encryptedPassword);
+    }
+}
