+++ src/main/resources/static/js/account/signup.js
