+++ src/main/resources/templates/account/login.html
@@ -1,0 +1,64 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <title>Login</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+    <link href="/css/account/login.css" rel="stylesheet"> <!-- 커스텀 CSS -->
+</head>
+<body class="bg-light">
+
+<div class="container d-flex justify-content-center align-items-center vh-100">
+    <div class="card shadow-lg p-4" style="width: 100%; max-width: 600px;">
+        <h2 class="text-center mb-4">로그인</h2>
+        <form th:action="@{/login}" th:object="${loginUserForm}" method="post">
+            <!-- 아이디 -->
+            <div class="form-group mb-3">
+                <label for="username" class="form-label">아이디</label>
+                <input type="email" id="username" name="username"
+                       class="form-control"
+                       th:field="*{email}"
+                       placeholder="example@example.com" required>
+            </div>
+            <!-- 비밀번호 -->
+            <div class="form-group mb-4">
+                <label for="password" class="form-label">비밀번호</label>
+                <input type="password" id="password" name="password"
+                       class="form-control"
+                       th:classappend="${errorMessage != null} ? 'is-invalid' : ''"
+                       placeholder="비밀번호를 입력하세요" required>
+                <div class="invalid-feedback" th:if="${errorMessage}">
+                    <span th:text="${errorMessage}"></span>
+                </div>
+            </div>
+            <!-- 로그인 버튼 -->
+            <button type="submit" class="btn btn-primary w-100">로그인</button>
+        </form>
+
+        <!-- 소셜 로그인 -->
+        <div class="text-center my-3">또는</div>
+        <div class="d-flex flex-column align-items-center gap-3">
+            <!-- 네이버 로그인 -->
+            <a href="/oauth2/authorization/naver" class="btn">
+                <img width="180" src="/images/login/naver_login.png" alt="네이버 로그인">
+            </a>
+            <!-- 카카오 로그인 -->
+            <a href="/oauth2/authorization/kakao" class="btn">
+                <img width="150" src="/images/kakao_login.png" alt="카카오 로그인">
+            </a>
+            <!-- 구글 로그인 -->
+            <a href="/oauth2/authorization/google" class="btn">
+                <img width="150" src="/images/google_login.png" alt="구글 로그인">
+            </a>
+        </div>
+
+        <!-- 회원가입 -->
+        <div class="text-center mt-4">
+            <p class="mb-0">아직 회원이 아니신가요?</p>
+            <a href="/signup" class="text-decoration-none">회원가입</a>
+        </div>
+    </div>
+</div>
+</body>
+</html>
