+++ src/main/resources/templates/account/signup.html
@@ -1,0 +1,119 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>회원가입</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css">
+    <link href="/css/account/signup.css" rel="stylesheet"> <!-- 커스텀 CSS -->
+    <!-- 모듈 스크립트 -->
+    <script type="module">
+        import { validateForm } from '/js/auth/signup.js';
+        validateForm();
+    </script>
+</head>
+<body>
+<div class="container mt-5">
+    <h2 class="text-center mb-4">회원가입</h2>
+    <form th:action="@{/signup}" th:object="${userSignUp}" method="post" class="needs-validation" novalidate>
+        <!-- 숨겨진 필드에 socialType 값 추가 -->
+        <input type="hidden" th:field="*{socialType}" th:value="${socialType}" />
+        <!-- 숨겨진 필드에 providerId 값 추가 -->
+        <input type="hidden" th:field="*{providerId}" th:value="${providerId}" />
+        <!-- 이름 -->
+        <div class="form-group">
+            <label for="name" class="form-label">이름</label>
+            <input type="text" id="name" th:field="*{name}"
+                   class="form-control"
+                   th:value="*{name}"
+                   th:readonly="${isSocial}"
+                   placeholder="이름을 입력하세요" required>
+            <div class="invalid-feedback">이름을 입력하세요.</div>
+        </div>
+        <!-- 닉네임 -->
+        <div class="form-group">
+            <label for="nickName" class="form-label">닉네임</label>
+            <input type="text" id="nickName" th:field="*{nickName}"
+                   class="form-control"
+                   placeholder="닉네임을 입력하세요" required>
+            <div class="invalid-feedback">닉네임을 입력하세요.</div>
+        </div>
+        <!-- 이메일 -->
+        <div class="form-group">
+            <label for="email" class="form-label">이메일</label>
+            <input type="email" id="email" th:field="*{email}"
+                   class="form-control"
+                   th:value="*{email}"
+                   th:readonly="${isSocial}"
+                   placeholder="example@example.com" required>
+            <div class="invalid-feedback">유효한 이메일을 입력하세요.</div>
+        </div>
+        <!-- 비밀번호 -->
+        <div class="form-group">
+            <label for="password" class="form-label">비밀번호</label>
+            <input type="password" id="password" th:field="*{password}"
+                   class="form-control"
+                   placeholder="비밀번호를 입력하세요" required>
+            <div class="invalid-feedback">비밀번호를 입력하세요.</div>
+        </div>
+        <!-- 전화번호 -->
+        <div class="form-group">
+            <label for="phoneNumber" class="form-label">전화번호</label>
+            <input type="text" id="phoneNumber" th:field="*{phoneNumber}"
+                   class="form-control"
+                   placeholder="010-1234-5678" required>
+            <div class="invalid-feedback">전화번호를 입력하세요.</div>
+        </div>
+        <!-- 주소 -->
+        <div class="form-group">
+            <label for="address-group" class="form-label">주소</label>
+            <div id="address-group" class="row g-3">
+                <div class="col-md-6">
+                    <input type="text" id="zipcode" th:field="*{zipcode}"
+                           class="form-control"
+                           placeholder="우편번호" required>
+                </div>
+                <div class="col-md-6">
+                    <input type="text" id="state" th:field="*{state}"
+                           class="form-control"
+                           placeholder="도/시" required>
+                </div>
+                <div class="col-md-6">
+                    <input type="text" id="city" th:field="*{city}"
+                           class="form-control"
+                           placeholder="시/군/구" required>
+                </div>
+                <div class="col-md-6">
+                    <input type="text" id="street" th:field="*{street}"
+                           class="form-control"
+                           placeholder="도로명/지번" required>
+                </div>
+                <div class="col-md-12">
+                    <input type="text" id="detailAddress" th:field="*{detailAddress}"
+                           class="form-control"
+                           placeholder="상세 주소" required>
+                </div>
+                <div class="col-md-12">
+                    <input type="text" id="addressName" th:field="*{addressName}"
+                           class="form-control"
+                           placeholder="주소 별칭" required>
+                </div>
+            </div>
+        </div>
+        <!-- Role 선택 -->
+        <div class="form-group">
+            <label for="role" class="form-label">권한</label>
+            <select id="role" th:field="*{role}" class="form-select" required>
+                <option value="" disabled selected>권한을 선택하세요</option>
+                <option value="ROLE_USER">사용자</option>
+                <option value="ROLE_SELLER">사장님</option>
+            </select>
+            <div class="invalid-feedback">권한을 선택하세요.</div>
+        </div>
+        <!-- 가입 버튼 -->
+        <div class="text-center">
+            <button type="submit" class="btn btn-primary">가입하기</button>
+        </div>
+    </form>
+</div>
+</body>
+</html>
